bugfix: render variant where priors table is empty
Previously, some old variants were loading as a blank page with the error message
"Uncaught TypeError: Cannot read properties of null (reading 'varType')"
eg, variant/61422

This was due to trying to read the varType column from the insilicopriors table when  no rows in that table contained
Variant_id = the variant in question.

This change makes the varType check conditional instead so the page can load.